AuthorshipDetector and Controller, FeatureTypes enum, init Result Parser from double to %.

+++ src/com/company/signature_feature/AverageSentenceComplexity.java
@@ -10,7 +10,7 @@ private FeatureDto featureDto;
 
   @Override
   public double featureCalculation(String text) {
-    return featureDto.getNumberOfPhrasesInText(text)
+    return (double) featureDto.getNumberOfPhrasesInText(text)
         / featureDto.getNumberOfSentencesInText(text);
   }
 }
